+++ server/src/formatter/formatterState.ts
@@ -1,4 +1,4 @@
-import {Node_Script} from '../compiler_parser/nodes';
+import {Node_Script} from '../compiler_parser/nodeObject';
 import {Position} from 'vscode-languageserver';
 import {TextEdit} from 'vscode-languageserver-types/lib/esm/main';
 import {TokenBase, TokenObject} from '../compiler_tokenizer/tokenObject';
