+++ server/src/inspector/analysisResolver.ts
@@ -1,6 +1,6 @@
 import * as lsp from 'vscode-languageserver/node';
 import {TokenObject, StringToken} from '../compiler_tokenizer/tokenObject';
-import {Node_Script} from '../compiler_parser/nodes';
+import {Node_Script} from '../compiler_parser/nodeObject';
 import {DelayedTask} from '../utils/delayedTask';
 import {PublishDiagnosticsParams} from 'vscode-languageserver-protocol';
 import {getGlobalSettings} from '../core/settings';
