+++ server/src/services/inlayHint.ts
@@ -2,7 +2,7 @@ import {SymbolGlobalScope, SymbolScope} from '../compiler_analyzer/symbolScope';
 import {TextLocation} from '../compiler_tokenizer/textLocation';
 import {isNodeClassOrInterface} from '../compiler_analyzer/symbolObject';
 import * as lsp from 'vscode-languageserver/node';
-import {NodeName} from '../compiler_parser/nodes';
+import {NodeName} from '../compiler_parser/nodeObject';
 import {stringifyResolvedType} from '../compiler_analyzer/symbolStringifier';
 
 export function provideInlayHint(globalScope: SymbolGlobalScope, location: TextLocation): lsp.InlayHint[] {
